Fix octree and migrate to cmake

+++ src/layout.cpp
@@ -175,7 +175,7 @@ void CLayout::load_positions(const fs::path & pathPositionsFile)
 {
     const auto positions = ReadFile(pathPositionsFile);
     #pragma omp parallel for
-    for (size_t i = 0; i < m_vecBodies.size(); ++i)
+    for (int i = 0; i < m_vecBodies.size(); ++i)
     {
         Vector3 pos(
             positions[i * 3 + 0],
@@ -200,7 +200,7 @@ void CLayout::load_weights(const fs::path & pathWeightsFile)
     }
 
     #pragma omp parallel for
-    for (size_t i = 0; i < m_vecBodies.size(); ++i)
+    for (int i = 0; i < m_vecBodies.size(); ++i)
     {
         m_vecBodies[i].set_mass(weights[i]);
     }
@@ -211,7 +211,7 @@ void CLayout::init_positions()
     const size_t maxBodyId = m_vecBodies.size();
 
     #pragma omp parallel for
-    for (size_t i = 0; i < m_vecBodies.size(); ++i)
+    for (int i = 0; i < m_vecBodies.size(); ++i)
     {
         auto& body = m_vecBodies[i];
         if (!body.get_position().is_initialized())
@@ -227,7 +227,7 @@ void CLayout::init_positions()
         auto& springs = body.get_springs();
         // init neighbours position:
         #pragma omp parallel for
-        for (size_t j = 0; j < springs.size(); ++j)
+        for (int j = 0; j < springs.size(); ++j)
         {
             if (!m_vecBodies[springs[j]].get_position().is_initialized())
             {
@@ -245,7 +245,7 @@ void CLayout::init_positions()
 void CLayout::init_weights()
 {
     #pragma omp parallel for
-    for (size_t i = 0; i < m_vecBodies.size(); ++i)
+    for (int i = 0; i < m_vecBodies.size(); ++i)
     {
         auto& body = m_vecBodies[i];
         body.set_mass(1 + (body.get_springs().size() + body.get_incoming_links()) / 3.0);
@@ -257,7 +257,7 @@ void CLayout::accumulate()
     m_Tree.insertBodies(m_vecBodies);
 
     #pragma omp parallel for
-    for (size_t i = 0; i < m_vecBodies.size(); ++i)
+    for (int i = 0; i < m_vecBodies.size(); ++i)
     {
         CBody& body = m_vecBodies[i];
         body.reset_force();
@@ -267,7 +267,7 @@ void CLayout::accumulate()
     }
 
     #pragma omp parallel for
-    for (size_t i = 0; i < m_vecBodies.size(); ++i)
+    for (int i = 0; i < m_vecBodies.size(); ++i)
     {
         updateSpringForce(m_vecBodies[i]);
     }
@@ -283,7 +283,7 @@ double CLayout::integrate()
     //dx should be private or defined inside loop
     //tx need to be reduction variable, or its value will be unpredictable.
     #pragma omp parallel for reduction(+:tx,ty,tz) private(dx,dy,dz)
-    for (size_t i = 0; i < m_vecBodies.size(); ++i)
+    for (int i = 0; i < m_vecBodies.size(); ++i)
     {
         CBody& body = m_vecBodies[i];
         const auto& force = body.get_force();
